requirements.txt
kafka-python>=2.0.2
confluent-kafka>=2.0.2
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
asyncio-mqtt>=0.13.0
websockets>=11.0.0
aiohttp>=3.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0 